+++ tests/test_services/test_nexus.py
@@ -6,8 +6,7 @@ import pytest
 from unittest.mock import Mock, patch
 from api.services.nexus import fetch_file_content
 from api.exceptions import AuthenticationIssueException, AuthorizationIssueException, ResourceNotFoundException
-from tests.fixtures.utils import load_content
-from tests.fixtures.nexus import morphology_content_url, access_token
+from tests.utils import load_content
 
 
 @patch("requests.get")